Change the way the trajectories are iteratively loaded
The point of this branch is to be the most memory efficient.
I've been running into memory issues with long trajectories and a long list of pairs.
Thus, this branch aims to fix that.
The first way is to give the MSIBI class the maximum number of frames read at once to
calculate the RDF.