Remove plotting; cast n_frames as int for range()
The plotting was just something to use while writing the initial code - it does not
belong in the repository. There will eventually be a directory full of example plotting
scripts to use.

In the new frame chunking scheme, n_frames is passed. This needs to be an int as it is used
in a range(). Thus, n_frames is xplicitly cast as int in the range() function.