Add select_pairs() function for selecting pairs
The current way of selecting pairs is to load in a list or pairs.
This is not ideal, as you must define these pairs before hand.
This function alleviates that issue.
Note that the exclusion calculation is the naive implementation, as it
makes a list of pairs to delete, and then uses numpy.delete() to delete
those rows from the numpy array.
I couldn't figure out how to use np.where() or np.apply_along_axis() to
get the same behavior.
Presumably, the numpy functions would be faster than the current
implementation.