Add functionality to support hoomd2
This involves writing new hoomd headers that get written when saving
a hoomd runscript. In MSIBI.optimize(), the hoomd version is deduced
with a try-except block.

In workers.py, the _hoomd_worker() now uses the correct executable
depending on the hoomd version, because hoomd1 was a standalone
executable, whereas hoomd2 is now a python module.