Update tests
A lot of the tests were dependent on the rdf length needing to be
specified with an extra bin, but mdtraj now does this as expected so
it's no longer need.

Also, use pytest instead of py.test for testing. I was getting weird
errors when using py.test: tests were failing because ```import hoomd```
was failing; using pytest instead of py.test seems to fix this issue.