Add tutorial for propane using similar Lennard-Jones format
Modify opt.py to reflect different file structures with the propane
rdfs and state names. Folder potentials not included because not
necessary for this tutorial.  Add plot.py and opt.out files to
allow results from the runs to be plotted on a graph. Lines 10-14 in
opt.py are kept as comments for similarity purposes ot the lj
tutorial, but they may be removed when necessary.