Propane Tutorial and Notebook Version (#31)
* Update README with propane

* Add tutorial for propane using similar Lennard-Jones format

Modify opt.py to reflect different file structures with the propane
rdfs and state names. Folder potentials not included because not
necessary for this tutorial.  Add plot.py and opt.out files to
allow results from the runs to be plotted on a graph. Lines 10-14 in
opt.py are kept as comments for similarity purposes ot the lj
tutorial, but they may be removed when necessary.

* Modify propane tutorial to include jupyter notebook edition

Divide the opt.py file for the python tutorial into sections.
Descriptions were included under each section to elaborate on the
purpose of that code section. Format follows standard jupyter
notebook protocol.

* Correct file directory structure by fixing file paths

* Cleanup python tutorial and fix README

* Update README with correct header and information

* Cleanup propane.ipynb and README files

* Modify README with correct citation for literature

* Remove generated animation and figure files from repo

* Remove setup.py